refactor: rename forget password and otp screen packages
