+++ identity_presentation/src/commonMain/kotlin/net/thechance/mena/identity/presentation/di/identityScreensModule.kt
@@ -1,7 +1,7 @@
 package net.thechance.mena.identity.presentation.di
 
-import net.thechance.mena.identity.presentation.screen.forget_password.ForgetPasswordScreenViewModel
-import net.thechance.mena.identity.presentation.screen.forget_password_otp.OTPScreenViewModel
+import net.thechance.mena.identity.presentation.screen.forgetPassword.ForgetPasswordScreenViewModel
+import net.thechance.mena.identity.presentation.screen.forgetPasswordOTP.OTPScreenViewModel
 import net.thechance.mena.identity.presentation.screen.login.LoginScreenModel
 import net.thechance.mena.identity.presentation.screen.register.RegisterScreenModel
 import org.koin.core.module.dsl.factoryOf
