+++ identity_presentation/src/commonMain/kotlin/net/thechance/mena/identity/presentation/screen/forgetPasswordOTP/OTPScreenUIEffect.kt
@@ -1,4 +1,4 @@
-package net.thechance.mena.identity.presentation.screen.forget_password_otp
+package net.thechance.mena.identity.presentation.screen.forgetPasswordOTP
 
 sealed class OTPScreenUIEffect {
     data object NavigateToResetPassword : OTPScreenUIEffect()
